Add the ability to download images
Previously only Pixiv was able to download images from a given Search.Post.SearchResult value. The 2 methods in Pixiv responsible for this functionality were pulled to base ABooru class, so any booru now has the ability to download images.

In addition to the above two methods,  boorus can also implement a method for downloading all of the images that are present in the supplied Search.Post.SearchResult value (for now only Pixiv implements this). Relevant flag and property were added for checking whether this booru functionality is present.

The downside of allowing all boorus to download posts is that there's no way for a booru to know the origin of the Search.Post.SearchResult, so for example Gelbooru instance can download posts retrieved by DanbooruDonmai instance. In theory this should work anyway since no data that is hard wired to a booru is used.